move user entity factor a bit up in test hierarchy
