+++ src/tests/fixtures/__init__.py
@@ -2,3 +2,4 @@ from tests.fixtures.docker import *
 from tests.fixtures.mongodb import *
 from tests.fixtures.api import *
 from tests.fixtures.rabbitmq import *
+from tests.fixtures.user_entity import UserEntityProtocol, user_entity_factory
