+++ src/tests/tests/domains/users/test_users_interactor_get.py
@@ -6,7 +6,7 @@ from innonymous.domains.users.errors import UsersError, UsersNotFoundError
 from innonymous.domains.users.interactors import UsersInteractor
 from innonymous.domains.users.entities import UserCredentialsEntity, UserEntity
 
-from tests.tests.domains.users.conftest import UserEntityProtocol
+from tests.conftest import UserEntityProtocol
 
 __all__ = ()
